+++ Projet_Gestion_Contacts/pom.xml
@@ -72,6 +72,7 @@
     </dependencies>
 
     <properties>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <maven.compiler.source>8</maven.compiler.source>
         <maven.compiler.target>8</maven.compiler.target>
     </properties>
@@ -91,7 +92,7 @@
                 <artifactId>javafx-maven-plugin</artifactId>
                 <version>0.0.5</version>
                 <configuration>
-                    <mainClass>View.App</mainClass>
+                    <mainClass>contact.App</mainClass>
                 </configuration>
             </plugin>
         </plugins>
