+++ Projet_Gestion_Contacts/src/main/java/contact/Launcher.java
@@ -1,4 +1,4 @@
-package View;
+package contact;
 
 public class Launcher {
 	public static void main(String[] args) {
