+++ Projet_Gestion_Contacts/src/main/java/view/HomeScreenController.java
@@ -2,6 +2,7 @@ package View;
 
 import java.io.IOException;
 
+import contact.App;
 import javafx.fxml.FXML;
 
 public class HomeScreenController {
@@ -10,6 +11,8 @@ public class HomeScreenController {
 		// Here we make use of our new method allowing us to change views inside the main Parent		
 		App.showView("Contact");
 	}
+	
+	@FXML
 	public void handleLaunchButton2() throws IOException {
 		// Here we make use of our new method allowing us to change views inside the main Parent		
 		App.showView("ListAllContact");
