--- Projet_Gestion_Contacts/src/main/resources/Contact.fxml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.DatePicker?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.AnchorPane?>
-
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="508.0" prefWidth="660.0" xmlns="http://javafx.com/javafx/15.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="projet.contact.view.ContactController">
-   <children>
-      <Label layoutX="180.0" layoutY="14.0" text="Modifier un contact" />
-      <Label layoutX="20.0" layoutY="44.0" text="FirstName" />
-      <Label layoutX="20.0" layoutY="88.0" text="LastName" />
-      <Label layoutX="19.0" layoutY="134.0" text="NickName" />
-      <Label layoutX="19.0" layoutY="182.0" text="PhoneFix" />
-      <Label layoutX="19.0" layoutY="219.0" text="MobilePhone" />
-      <Label layoutX="20.0" layoutY="262.0" text="Birthday" />
-      <Label layoutX="20.0" layoutY="302.0" text="Mail" />
-      <Label layoutX="19.0" layoutY="339.0" text="Adress" />
-      <Label layoutX="20.0" layoutY="380.0" text="Website" />
-      <TextField fx:id="firstName" layoutX="105.0" layoutY="40.0" prefHeight="26.0" prefWidth="254.0" />
-      <TextField fx:id="lastName" layoutX="105.0" layoutY="83.0" prefHeight="26.0" prefWidth="254.0" />
-      <TextField fx:id="nickName" layoutX="105.0" layoutY="130.0" prefHeight="26.0" prefWidth="254.0" />
-      <TextField fx:id="phoneFix" layoutX="105.0" layoutY="177.0" prefHeight="26.0" prefWidth="254.0" />
-      <TextField fx:id="mobilePhone" layoutX="105.0" layoutY="215.0" prefHeight="26.0" prefWidth="254.0" />
-      <TextField fx:id="mail" layoutX="105.0" layoutY="375.0" prefHeight="26.0" prefWidth="254.0" />
-      <TextField fx:id="adress" layoutX="105.0" layoutY="335.0" prefHeight="26.0" prefWidth="254.0" />
-      <TextField fx:id="website" layoutX="105.0" layoutY="297.0" prefHeight="26.0" prefWidth="254.0" />
-      <Button layoutX="451.0" layoutY="461.0" mnemonicParsing="false" onAction="#handleSaveButton" text="Enregistrer" />
-      <Button layoutX="547.0" layoutY="461.0" mnemonicParsing="false" onAction="#handleDeleteButton" text="Supprimer" />
-      <DatePicker layoutX="105.0" layoutY="258.0" />
-      <Label layoutX="18.0" layoutY="416.0" text="City" />
-      <Label layoutX="16.0" layoutY="453.0" text="Country" />
-      <TextField fx:id="mail1" layoutX="104.0" layoutY="448.0" prefHeight="26.0" prefWidth="254.0" />
-      <TextField fx:id="mail2" layoutX="105.0" layoutY="411.0" prefHeight="26.0" prefWidth="254.0" />
-   </children>
-</AnchorPane>
