--- Projet_Gestion_Contacts/src/main/resources/HomeScreen.fxml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.String?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-<?import javafx.scene.text.Text?>
-<?import javafx.scene.text.TextFlow?>
-
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/15.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="View.HomeScreenController">
-   <children>
-      <Button id="homeButton1" layoutX="146.0" layoutY="121.0" mnemonicParsing="false" onAction="#handleLaunchButton" prefHeight="95.0" prefWidth="122.0" text="Ajouter un contact" />
-      <Button id="homeButton2" layoutX="313.0" layoutY="124.0" mnemonicParsing="false" onAction="#handleLaunchButton" prefHeight="95.0" prefWidth="122.0" text="Afficher les contacts" />
-      <Text layoutX="195.0" layoutY="51.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Bienvenue dans votre carnet de contacts" wrappingWidth="219.3570556640625" />
-   </children>
-</AnchorPane>
