+++ Projet_Gestion_Contacts/src/main/resources/View/HomeScreen.fxml
@@ -9,8 +9,8 @@
 
 <AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/15.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="View.HomeScreenController">
    <children>
-      <Button id="homeButton1" layoutX="146.0" layoutY="121.0" mnemonicParsing="false" onAction="#handleLaunchButton" prefHeight="95.0" prefWidth="122.0" text="Ajouter un contact" />
-      <Button id="homeButton2" layoutX="313.0" layoutY="124.0" mnemonicParsing="false" onAction="#handleLaunchButton" prefHeight="95.0" prefWidth="122.0" text="Afficher les contacts" />
+      <Button id="homeButton1" layoutX="146.0" layoutY="121.0" mnemonicParsing="false" onAction="#handleLaunchButton1" prefHeight="95.0" prefWidth="122.0" text="Ajouter un contact" />
+      <Button id="homeButton2" layoutX="313.0" layoutY="124.0" mnemonicParsing="false" onAction="#handleLaunchButton2" prefHeight="95.0" prefWidth="122.0" text="Afficher les contacts" />
       <Text layoutX="195.0" layoutY="51.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Bienvenue dans votre carnet de contacts" wrappingWidth="219.3570556640625" />
    </children>
 </AnchorPane>
